Set mower to idle state upon reaching daily limit (#181)
In accordance with the GARDENA Operator's Manual, distinct mower models have varying maximum work capacity that either is defined by a minimum daily standby time or a maximum cutting time within a day.

Previously, upon reaching the daily limit, the mower's state was set to error, which implied a need for manual intervention to rectify the situation. However, the daily limit reached, is only a temporary state and resets automatically at midnight and eliminating the need for manual intervention.